requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
# Core LangChain Stack - Pinned for compatibility
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.25
langchain-groq==0.2.1
langsmith==0.1.145
langchain-google-genai==2.0.1
langchain-text-splitters==0.3.3

# RAG Stack
haystack-ai>=2.11.0
pinecone-haystack==5.4.0
# Use newer pinecone that has the functionality built-in
pinecone>=5.4.1
supabase==2.5.1

# Utils
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
requests==2.32.3
prometheus-client==0.21.0
sentry-sdk==2.18.0
python-multipart==0.0.9
websockets==12.0

# Search/Google Clients
azure-cognitiveservices-search-websearch==2.0.0
google-api-python-client==2.151.0

# ML/Vision - Updated for Python 3.13 compatibility
sentence-transformers>=3.1.1
transformers>=4.46.3
torch>=2.6.0
accelerate>=1.1.1
pymupdf>=1.24.10
pdf2image==1.17.0
pytesseract==0.3.13
Pillow>=11.0.0
unstructured>=0.15.0
opencv-python-headless>=4.10.0.84
scikit-learn>=1.5.2

# Misc
redis==5.2.0
slowapi==0.1.9
orjson==3.10.7
pytest==8.3.3
